requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
asttokens==3.0.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
emoji==2.14.1
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.8
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lancedb==0.19.0
langchain==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.10
libmagic==1.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.27.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.61.1
opt-einsum==3.3.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
paddlepaddle==2.6.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pylance==0.23.0
pyparsing==3.2.1
pypdf==5.3.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.20
unstructured-client==0.30.3
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
